Remove census processor and all multi processor switching
Only one processor will be needed so this can be invoked directly
