+++ tests/test_data.py
@@ -1,29 +1,3 @@
-ce_census_survey = '''
-{
-  "tx_id": "0f534ffc-9442-414c-b39f-a756b4adc6cb",
-  "type": "uk.gov.ons.edc.eq:surveyresponse",
-  "origin": "uk.gov.ons.edc.eq",
-  "version": "0.0.1",
-  "survey_id": "0",
-  "collection": {
-    "instrument_id": "hh2016",
-    "exercise_sid": "hfjdskf",
-    "period": "0616"
-  },
-  "submitted_at": "2016-03-12T10:39:40Z",
-  "metadata": {
-    "user_id": "789473423",
-    "ru_ref": "1234570071A"
-  },
-  "data": {
-    "1": "2",
-    "2": "4",
-    "3": "2",
-    "4": "Y"
-  }
-}
-'''
-
 common_software_survey = '''
 {
   "tx_id": "0f534ffc-9442-414c-b39f-a756b4adc6cb",
